rust: rename `platdev` to `platform`.
The module contains not only devices but also drivers, so reflect that
in the name. It is also meant to be consistent with other buses like
amba, pci, etc.

This is in preparation for converting it to use the common `driver` to
help implementing driver support.
