+++ rust/kernel/platform.rs
